[REFACTOR] components, containers 구조로 변경
각 폴더 아래에는 기능별로 서브폴더 존재
