+++ client/src/App.js
@@ -3,7 +3,8 @@ import ReactDOM from 'react-dom';
 import { BrowserRouter as Router, Switch, Route, Link } from 'react-router-dom';
 import { createGlobalStyle } from 'styled-components';
 
-import { LoginContainer, GitHubCallback } from './auth';
+import LoginContainer from '@/containers/login';
+import GitHubCallback from '@/components/login/github';
 const IssueContainer = null; // import IssueContainer from './issue';
 
 const GlobalStyle = createGlobalStyle`
@@ -23,6 +24,7 @@ const App = () => {
       <GlobalStyle />
       <Switch>
         <Route path="/" exact component={!token ? LoginContainer : IssueContainer} />
+        <Route path="/issues" exact component={IssueContainer} />
         <Route path="/users/github/callback" render={props => <GitHubCallback {...props} cb={setToken} />} />
       </Switch>
     </>
