--- client/src/auth/index.js
@@ -1,4 +1,0 @@
-import LoginContainer from './login';
-import GitHubCallback from './github';
-
-export { LoginContainer, GitHubCallback };
